scout_backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
djangorestframework-simplejwt==5.5.1
Pillow==11.3.0
PyJWT==2.10.1
